Changed cn_rx.py script.
Now cn_rx.py takes an argument which is the number of antennas in the
correlator you wish to run. This should be the full number of antennas
even if you do not have all the hardware etc...
This gets rid of the need to change the ants in cn_rx everytime we want
to run our 16 or 32 antenna correlators.
    eg. python cn_rx.py 32 (for 32 antenna correlator).

git-svn-id: https://casper.berkeley.edu/svn/trunk/projects/paper@3432 577e11ce-4b3a-42a0-b907-5d4e03b181e5