Changes to corr_adc_ampl.py
Looking forward to changes on the ibob, which will send over the mean
values of the adc samples so that we can correctly calculate the rms
values coming in from the input mux.
The changes here allude to that. They snap the mean values also and
calculate the rms values with the formula:
rms = sqrt((1/n)(sum(x^2)) - mean(x)^2)

git-svn-id: https://casper.berkeley.edu/svn/trunk/projects/paper@3442 577e11ce-4b3a-42a0-b907-5d4e03b181e5