Bug fix for Vacc sync up.
When Vacc resync ran, there was a shift in the ordering
of the baselines that occured. This was due to a delay
mismatch between the mcnt and data. This bug was fixed
by adding a delay of 4 to the data path.
Note that previously, this delay of four was incorporated into
the Vacc TVG which has been removed for timing reasons.

git-svn-id: https://casper.berkeley.edu/svn/trunk/projects/paper@3458 577e11ce-4b3a-42a0-b907-5d4e03b181e5