Improve adc power level logic
Improve the ADC power level logic with these changes:

* Measure input levels after input mux rather than always measuring ADC levels.

* Include sum of samples as well as sum of squares.  This allows the mean (aka
  DC offset, mostly an ADC artifact) and standard deviation to be properly
  computed.  Using just the sum of squares is not so robust as it assumes that
  the mean is 0, which is rarely ever the case even for terminated ADC inputs
  due to ADC offset errors.

* Fixes a small bug in the summers so that they now really do accumulate 65536
  samples rather than 65535 samples as had been the case.

The sum of samples gets passed over the XAUI link along with antenna number and
sum of squares all in one 64-bit value.  The upper 24 bits are the (signed!)
sum of samples.  The lower 32 bits are the (unsigned) sum of squares.  The 8
bits in between are the antenna number.