Add FFT preshift functionality
The preshift is controlled by bits 26 and 27 of the ctrl register.  The
preshift allows for 0, 1, 2, or 3 bits of DOWN shifting after the polyphase
filter but before the FFT.  This is intended to prevent saturation in the first
one or two stages of the FFT in the event of excessively strong input signals
(e.g. RFI).