Add conf keys n_ants_per_feng and xeng_chan_mode
n_ants_per_feng (an integer) specifies the number of antennas per F
engine.  I prefer to think in terms of inputs, but this is in terms of
antennas to match the units of n_ants.

xeng_chan_mode (and integer) specifies the nature of the channels that
an X engine processes.  xeng_chan_mode=0 means strided (aka interleaved)
and xeng_chan_mode=1 means contiguous.