Support new name for sync time key
The time of the most recent sync used to be stored in memcache using the
key "mcount_initialize_time".  It is now stored using the key
"roachf_init_time".  cn_rx.py now tries to fetch the time using the new
key, but if that fails, it falls back to getting it using the old key.